requirements.txt
google-genai
streamlit
tensorflow
tf-keras
pillow
numpy
